Working advanced mode integration tests
In order to get the integration tests working, I had to add a --wipe-all-configs flag
to keybaseca. This command is run to wipe all kssh client config files from KBFS
in order to ensure that the multiple runs of tests do not interfere with each other.
As part of that, I moved all of the KBFS code to shared/ so that keybaseca and kssh
share a simple library to interact with KBFS via `keybase fs` commands.

I also added better checking for the integration tests via the integrationTestUtils.py
script which reads the output from the integration tests and ensures that the strings
"kssh passed test N" show up in the output. Previously, I relied on non-zero exit codes
to signal failure. Now a test fails if there is a non-zero exit code or if the correct
string is not printed.