Write the client config into each listed team
kssh now deduplicates client configs based off of the botname. This is needed to deal with this config file

```
use_subteams_as_principals: true
teams:
- team.ssh.prod
- team.ssh.staging
```

in the case where a user has access to staging but not prod. They would not have the ability to read
the KBFS file at /keybase/team/team.ssh.prod/kssh-client.config and thus kssh would not work correctly.
The deduplication based off of botname is necessary in order to ensure that a user with access to both
teams does not have to use the --team flag since there really is only one bot.