Add script to automatically create and configure accounts for use with the integration tests
tests/configure_accounts.py walks the user through creating two new accounts, prompts the user
for the username and paperkeys, and then automatically creates the necessary shared teams between
the two accounts. It then writes this data to the env.sh files in the test directories. This
gets rid of the need for the env.sh.example files. This script is called by integrationTest.sh
if the env.sh files do not already exist.

In addition, I moved tests/{multi,single}-environment/bot-entrypoint.sh to a shared directory
so as to reduce the amount of duplication in the tests directory.

I attempted to automate the actual creating of the accounts, but the `keybase signup` command
uses the golang.org/x/crypto/ssh/terminal.ReadLine method to read input which assumes that the
user is using a terminal. Because of this, it is not possible to naively just pipe stdin to the
signup command. It would be possible to do using a terminal emulator like pyte (https://pyte.readthedocs.io/en/latest/)
but that adds a significant amount of complexity to the set up script.

As part of this, I also migrated the multi-environment tests to exercise the code contained in
the david/client-config-everywhere branch. This is done by having the first listed team in the
keybaseca.config file be a team that the KSSH user is not in.