Test the CA bot with multiple configs
This creates the necessary infrastructure to easily test the CA bot with multiple configs. The CA bot
container now contains a simple python webserver that will kill the CA bot and reload it with a specified
env file. Currently, this is only used with two different env files (one for the majority of the tests and
one for testing the audit logs writing to the normal FS) but it could be easily expanded to N env files