Working tests for writing the client config everywhere
Thanks to the new test framework that allows writing integration tests for different server
configs, this adds tests that exercise the code in this branch. It adds env-3 which tests
that kssh works properly when the kssh user is not in the first team listed in the config
file. I also added env-4 which tests that keybaseca does not sign requests that come in
outside of the configured teams.