Add ability to set a default SSH user for kssh
This feature is useful when using jumpboxes combined with ssh configs. For example, imagine you have an SSH
config that translates to this command:

```
ssh -J jump.example.com server.internal
```

When run via kssh, you want it to use the `developer` user but when run via ssh you want it to use your
username as the user (the default when no user is specified). Prior to this change, that was not possible.
Now, kssh does two new things:

1. kssh --set-default-user
	It is now possible to set a default SSH user to use with kssh. This is implemented via an SSH config
	file that applies a User directive to all hosts. This config file then inherits from the default ssh
	config file at ~/.ssh/config. This makes it possible to set a default user without modifying the
	user's ssh config file.
2. kssh now adds to the SSH agent by default
	Now by default kssh adds to the running ssh agent. This is necessary in order for jumpboxes to work
	since they rely on ssh agent forwarding.

As part of this, I also had to tweak the tests to make all of this work properly.