Add information on how to use scp and rsync with kssh and default users
This change makes it so kssh --provision creates the kssh ssh config file (the one that specifies a default ssh user). It then
warns the user if they have a default user configured that they need to see the docs on how to use scp and rsync with the
default user. The docs explain the problem, how to solve it, and provide aliases to do it easily.