Fix gpg signature verification
This fixes gpg signature verification to not rely on the key servers (without just blindly trusting
the TLS connection to Keybase). The PGP key is downloaded from keybase and it is verified that the
key fingerprint matches the correct fingerprint. Relying on the key servers caused flakey builds in
some cases.