Working integration tests for two man mode!
keybaseca now supports signature requests that specify a requested realm. If the requested realm is one of a configured list
of two man realms, it requires confirmation from another party. A list of approvers is defined in the keybaseca config. The
approvers are tagged with a request for their approval. They approve a request by reacting with a thumbs-up emoji.

kssh now has a `kssh --request-realm foo` flag.

Integration tests are done with a bit of python code that reacts to messages with a thumbs-up emoji.

TODO:
* All the TODOs marked in the code
* More integration tests
* Have kssh track the current principals in a key and decide whether or not to renew (also good for debug logging)