Fix security vulnerability in WIP two-man mode
This makes it so it is not possible for the same person to double-approve a message. It used to be possible
to double-approve a message by reacting, unreacting, and reacting again. This is now prevented by keeping a
list of approvers and checking whether someone is double approving.

Also re-enables all the other integration tests that were disabled for testing purposes.